backend/requirements.txt
# =========================
# API core
# =========================
fastapi==0.119.0
uvicorn==0.37.0
python-dotenv==1.1.1
pymongo==4.15.3

# =========================
# Auth / Security
# =========================
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator


# =========================
# Pydantic v2
# =========================
pydantic==2.12.0
pydantic-settings==2.2.1

# =========================
# ML / RAG stack (CPU only)
# =========================
numpy==1.26.4
torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

openai-whisper
sentence-transformers==5.1.1
transformers==4.41.2
scikit-learn==1.7.2
scipy==1.16.2

# =========================
# NLP
# =========================
nltk
regex==2025.9.18

# =========================
# Utils
# =========================
requests==2.32.5
tqdm==4.67.1

# =========================
# Gemini client
# =========================
google-genai

python-multipart